feat(Jest-runner): support jest configuration (#25)
Support configurable options using the `jest` property in your stryker config:

``` js
jest: {
    project: 'react', // or 'default'
    config: require('path/to/your/custom/jestConfig.js') 
}
```

When neither of the options are specified, the jest config in your `package.json` file is used.

To support reading jest configuration, we are now using a higher level jest api, the `jest-cli`, to start the tests. In order to keep it in a single process, the `run-in-band` option is provided, forcing jest to run inside the parent process.